fix: merge global resources of i18n custom block
closes #805
