+++ packages/vue-i18n-core/src/i18n.ts
@@ -33,8 +33,8 @@ import { defineMixin as defineMixinBridge } from './mixins/bridge'
 import { enableDevTools, addTimelineEvent } from './devtools'
 import {
   isLegacyVueI18n,
-  getLocaleMessages,
-  getComponentOptions
+  getComponentOptions,
+  adjustI18nResources
 } from './utils'
 
 import type { ComponentInternalInstance, App } from 'vue'
@@ -865,47 +865,6 @@ function getGlobalComposer(i18n: I18n): Composer {
       ? (i18n.global as unknown as Composer)
       : (i18n.global as unknown as VueI18nInternal).__composer
     : (i18n.global as unknown as Composer)
-}
-
-function adjustI18nResources(
-  global: Composer,
-  options: UseI18nOptions,
-  componentOptions: any // eslint-disable-line @typescript-eslint/no-explicit-any
-): void {
-  let messages = isObject(options.messages) ? options.messages : {}
-  if ('__i18nGlobal' in componentOptions) {
-    messages = getLocaleMessages(global.locale.value as Locale, {
-      messages,
-      __i18n: componentOptions.__i18nGlobal
-    })
-  }
-  // merge locale messages
-  const locales = Object.keys(messages)
-  if (locales.length) {
-    locales.forEach(locale => {
-      global.mergeLocaleMessage(locale, messages[locale])
-    })
-  }
-  if (!__LITE__) {
-    // merge datetime formats
-    if (isObject(options.datetimeFormats)) {
-      const locales = Object.keys(options.datetimeFormats)
-      if (locales.length) {
-        locales.forEach(locale => {
-          global.mergeDateTimeFormat(locale, options.datetimeFormats![locale])
-        })
-      }
-    }
-    // merge number formats
-    if (isObject(options.numberFormats)) {
-      const locales = Object.keys(options.numberFormats)
-      if (locales.length) {
-        locales.forEach(locale => {
-          global.mergeNumberFormat(locale, options.numberFormats![locale])
-        })
-      }
-    }
-  }
 }
 
 function getComposer(
